requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.5.0,<6.0.0
Pillow>=10.0.0,<11.0.0  # 用于处理图像上传 